[Rename] client/benchmark
This PR renames Elasticsearch to OpenSearch in the client/benchmark module.
