+++ client/benchmark/build.gradle
@@ -17,16 +17,16 @@
  * under the License.
  */
 
-apply plugin: 'elasticsearch.build'
+apply plugin: 'opensearch.build'
 apply plugin: 'application'
 
-group = 'org.elasticsearch.client'
+group = 'org.opensearch.client'
 
 // Not published so no need to assemble
 assemble.enabled = true
 
 archivesBaseName = 'client-benchmarks'
-mainClassName = 'org.elasticsearch.client.benchmark.BenchmarkMain'
+mainClassName = 'org.opensearch.client.benchmark.BenchmarkMain'
 
 // never try to invoke tests on the benchmark project - there aren't any
 test.enabled = false
@@ -35,7 +35,7 @@ dependencies {
   api 'org.apache.commons:commons-math3:3.2'
 
   api project(":client:rest")
-  // bottleneck should be the client, not Elasticsearch
+  // bottleneck should be the client, not OpenSearch
   api project(path: ':client:client-benchmark-noop-api-plugin')
   // for transport client
   api project(":server")
