+++ client/benchmark/src/main/java/org/opensearch/client/benchmark/metrics/Metrics.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.client.benchmark.metrics;
+package org.opensearch.client.benchmark.metrics;
 
 public final class Metrics {
     public final String operation;
