+++ client/benchmark/src/main/java/org/opensearch/client/benchmark/rest/RestClientBenchmark.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.client.benchmark.rest;
+package org.opensearch.client.benchmark.rest;
 
 import org.apache.http.HttpHeaders;
 import org.apache.http.HttpHost;
@@ -26,9 +26,9 @@ import org.elasticsearch.OpenSearchException;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.RestClient;
-import org.elasticsearch.client.benchmark.AbstractBenchmark;
-import org.elasticsearch.client.benchmark.ops.bulk.BulkRequestExecutor;
-import org.elasticsearch.client.benchmark.ops.search.SearchRequestExecutor;
+import org.opensearch.client.benchmark.AbstractBenchmark;
+import org.opensearch.client.benchmark.ops.bulk.BulkRequestExecutor;
+import org.opensearch.client.benchmark.ops.search.SearchRequestExecutor;
 
 import java.io.IOException;
 import java.util.Collections;
